feat(api): Use new Parser and Generators
