+++ src/api/src/main.py
@@ -7,6 +7,7 @@ from typing import Optional
 from zipfile import ZipFile
 
 from aiohttp import ClientResponseError
+from bs4 import BeautifulSoup
 from eliot import start_action
 from fastapi import FastAPI, HTTPException, Request
 from fastapi.responses import (
@@ -19,6 +20,7 @@ from fastapi.staticfiles import StaticFiles
 
 from create_book import (
     EPUBGenerator,
+    PDFGenerator,
     StoryNotFoundError,
     WattpadError,
     fetch_cookies,
@@ -26,10 +28,10 @@ from create_book import (
     fetch_story,
     fetch_story_content_zip,
     fetch_story_from_partId,
-    generate_clean_part_html,
     logger,
     slugify,
 )
+from create_book.parser import clean_tree, download_tree_images
 
 app = FastAPI()
 BUILD_PATH = Path(__file__).parent / "build"
@@ -73,7 +75,7 @@ app.add_middleware(RequestCancelledMiddleware)
 
 
 class DownloadFormat(Enum):
-    # pdf = "pdf"
+    pdf = "pdf"
     epub = "epub"
 
 
@@ -169,30 +171,36 @@ async def handle_download(
         if not cover_data:
             raise HTTPException(status_code=422)
 
-        match format:
-            case DownloadFormat.epub:
-                book = EPUBGenerator(metadata, cover_data)
-                media_type = "application/epub+zip"
-            # case DownloadFormat.pdf:
-            #     book = PDFGenerator(metadata, cover_data)
-            #     media_type = "application/pdf"
-
-        logger.info(f"Retrieved story metadata and cover ({story_id=})")
-
         story_zip = await fetch_story_content_zip(story_id, cookies)
         archive = ZipFile(story_zip, "r")
 
-        part_contents = [
-            generate_clean_part_html(
-                part, archive.read(str(part["id"])).decode("utf-8")
+        part_trees: list[BeautifulSoup] = [
+            clean_tree(
+                part["title"], part["id"], archive.read(str(part["id"])).decode("utf-8")
             )
             for part in metadata["parts"]
         ]
 
-        async for title in book.add_chapters(
-            part_contents, download_images=download_images
-        ):
-            ...
+        if download_images:
+            images = [await download_tree_images(tree) for tree in part_trees]
+
+        match format:
+            case DownloadFormat.epub:
+                book = EPUBGenerator(metadata, part_trees, cover_data, images)
+                media_type = "application/epub+zip"
+            case DownloadFormat.pdf:
+                author_image = await fetch_image(
+                    metadata["user"]["avatar"].replace("-256-", "-512-")
+                )
+                if not author_image:
+                    raise HTTPException(status_code=422)
+
+                book = PDFGenerator(
+                    metadata, part_trees, cover_data, images, author_image
+                )
+                media_type = "application/pdf"
+
+        logger.info(f"Retrieved story metadata and cover ({story_id=})")
 
         book_buffer = book.dump()
